wip(Proposals): Added button to proposalcard
