+++ components/ProposalCard.tsx
@@ -1,194 +1,417 @@
-import { FC, useContext, useState } from "react";
-import { AppStateContext, contractStorage, tezosState } from "../context/state";
-import { proposal, proposalContent, status } from "../types/display";
-import { adaptiveTime, countdown } from "../utils/adaptiveTime";
-import useIsOwner from "../utils/useIsOwner";
-import { signers } from "../versioned/apis";
-import ContractLoader from "./contractLoader";
+import { InfoCircledIcon, TriangleDownIcon } from "@radix-ui/react-icons";
+import { useState } from "react";
+import { proposalContent } from "../types/display";
+import Alias from "./Alias";
+import Tooltip from "./Tooltip";
 
-function renderContent(
-  x: proposalContent,
-  state: tezosState,
-  address: string,
-  contract: contractStorage
-): string {
-  if ("transfer" in x) {
-    return `${x.transfer.amount} mutez to ${
-      state.aliases[x.transfer.destination] || x.transfer.destination
-    }`;
-  }
-  if ("executeLambda" in x) {
-    return `Execute Lambda(${x.executeLambda.metadata})`;
-  }
-  if ("execute" in x) {
-    return `Execute (${x.execute})`;
-  }
-  if ("adjustEffectivePeriod" in x) {
-    return `Adjust effective period:  (${adaptiveTime(
-      x.adjustEffectivePeriod.toString()
-    )})`;
-  }
-  if ("addOwners" in x) {
-    return `Add [${x.addOwners.join(", ")}] to validators`;
-  }
-  if ("removeOwners" in x) {
-    return `Remove [${x.removeOwners.join(", ")}] from validators`;
-  }
-  if ("changeThreshold" in x) {
-    return `Change threshold from ${contract.threshold} to ${x.changeThreshold}`;
-  }
-  let _: never = x;
-  return "Not supported";
-}
+type data = {
+  label: undefined | string;
+  metadata: undefined | string;
+  amount: undefined | string;
+  addresses: undefined | string[];
+  entrypoints: undefined | string;
+  params: undefined | string;
+};
 
-function getState(t: proposal): status {
-  return t.status;
-}
+const renderProposalContent = (content: proposalContent, i: number) => {
+  let data: data = {
+    label: undefined,
+    metadata: undefined,
+    amount: undefined,
+    addresses: undefined,
+    entrypoints: undefined,
+    params: undefined,
+  };
 
-const Card: FC<{
-  prop: { og: any; ui: proposal };
-  address: string;
-  id: number;
-  signable: boolean;
-  contract: contractStorage;
-  setCloseModal?: (arg: boolean | undefined) => void;
-}> = ({ contract, prop, address, id, signable, setCloseModal = () => {} }) => {
-  const isOwner = useIsOwner();
-  let state = useContext(AppStateContext)!;
-  let [loading, setLoading] = useState(false);
-  function resolvable(
-    signatures: { signer: string; result: boolean }[]
-  ): boolean {
-    let pro = signatures.filter(x => x.result).length >= contract.threshold;
-    let against = signatures.filter(x => !x.result).length > contract.threshold;
-    return pro || against;
+  if ("changeThreshold" in content) {
+    data = {
+      ...data,
+      label: "Update threshold",
+      params: content.changeThreshold.toString(),
+    };
+  } else if ("adjustEffectivePeriod" in content) {
+    data = {
+      ...data,
+      label: "Update proposal duration",
+      params: content.adjustEffectivePeriod.toString(),
+    };
+  } else if ("addOwners" in content) {
+    data = {
+      ...data,
+      label: `Add signer${content.addOwners.length > 1 ? "s" : ""}`,
+      addresses: content.addOwners,
+    };
+  } else if ("removeOwners" in content) {
+    data = {
+      ...data,
+      label: `Remove signer${content.removeOwners.length > 1 ? "s" : ""}`,
+      addresses: content.removeOwners,
+    };
+  } else if ("transfer" in content) {
+    data = {
+      ...data,
+      label: "Transfer",
+      addresses: [content.transfer.destination],
+      amount: content.transfer.amount.toString(),
+    };
+  } else if ("execute" in content) {
+    data = {
+      ...data,
+      label: "Execute",
+      metadata: content.execute,
+    };
+  } else if ("executeLambda" in content) {
+    const metadata = JSON.parse(content.executeLambda.metadata ?? "{}");
+
+    if (!metadata?.meta?.includes("contract_addr")) {
+      data = {
+        ...data,
+        label: "Execute lambda",
+        metadata:
+          metadata.meta === "No meta supplied" ? undefined : metadata.meta,
+      };
+    } else {
+      const contractData = JSON.parse(metadata.meta);
+      const endpoint = contractData.entrypoint;
+      const arg = contractData.payload;
+
+      data = {
+        label: "Execute contract",
+        metadata: contractData.meta,
+        amount: contractData.mutez_amount,
+        addresses: [contractData.contract_addr],
+        entrypoints: endpoint,
+        params: JSON.stringify(arg),
+      };
+    }
   }
+
   return (
-    <div className="rounded bg-zinc-800 px-6 py-4">
-      <div>
-        <p className="font-bold text-white md:inline-block">Status: </p>
-        <p className="md:text-md text-sm font-bold text-white md:inline-block">
-          {getState(prop.ui)}
+    <div
+      key={i}
+      className="after:content[''] relative grid grid-cols-3 gap-4 after:absolute after:left-0 after:right-0 after:-bottom-2 after:h-px after:bg-zinc-500 lg:grid-cols-6 lg:after:hidden"
+    >
+      <span
+        className={`${!data.label ? "text-zinc-500" : ""} justify-self-start`}
+      >
+        <p className="text-zinc-500 lg:hidden">Function</p>
+        {data.label ?? "-"}
+      </span>
+      <span
+        className={`${
+          !data.metadata ? "text-zinc-500" : ""
+        } w-full justify-self-center text-center lg:w-auto lg:justify-self-start lg:text-left`}
+      >
+        <p className="flex text-zinc-500 lg:hidden">
+          Metadata
+          <Tooltip text="Metadata is user defined. It may not reflect on behavior of lambda">
+            <InfoCircledIcon className="ml-2 h-4 w-4" />
+          </Tooltip>
         </p>
-      </div>
-      {"effective_period" in (contract ?? {}) && (
-        <div>
-          <p className="font-bold text-white md:inline-block">Expires in: </p>
-          <p className="md:text-md text-sm font-bold text-white md:inline-block">
-            {countdown(contract.effective_period, prop.ui.timestamp)}
-          </p>
-        </div>
+        {data.metadata ?? "-"}
+      </span>
+      <span
+        className={`${
+          !data.amount ? "text-zinc-500" : ""
+        } justify-self-end text-right lg:justify-self-center`}
+      >
+        <p className="text-zinc-500 lg:hidden">Amount</p>
+        {!data.amount ? "-" : `${data.amount} mutez`}
+      </span>
+      {!data.addresses ? (
+        <span className="justify-self-start text-zinc-500 lg:justify-self-center">
+          <p className="text-zinc-500 lg:hidden">Address</p>-
+        </span>
+      ) : data.addresses.length === 1 ? (
+        <span className="justify-self-start lg:justify-self-center">
+          <p className="text-zinc-500 lg:hidden">Address</p>
+          <Alias address={data.addresses[0]} />
+        </span>
+      ) : (
+        <ul className="justify-self-start lg:justify-self-center">
+          <li className="text-zinc-500 lg:hidden">Addresses:</li>
+          {data.addresses.map((address, i) => (
+            <li key={i}>
+              <Alias address={address} />
+            </li>
+          ))}
+        </ul>
       )}
-      <div>
-        <p className="font-bold text-white md:inline-block">Proposed by: </p>
-        <p className="md:text-md text-sm font-bold text-white md:inline-block">
-          {state.aliases[prop.ui.author] || prop.ui.author}
-        </p>
-      </div>
+      <span
+        className={`${
+          !data.entrypoints ? "text-zinc-500" : ""
+        } w-full justify-self-center text-center lg:w-auto lg:justify-self-end`}
+      >
+        <p className="text-zinc-500 lg:hidden">Entrypoint</p>
+        {data.entrypoints ?? "-"}
+      </span>
+      <span
+        className={`${
+          !data.params ? "text-zinc-500" : ""
+        } justify-self-end text-right`}
+      >
+        <p className="text-zinc-500 lg:hidden">Params</p>
+        {data.params ?? "-"}
+      </span>
+    </div>
+  );
+};
 
-      {("Executed" === prop.ui.status || "Rejected" === prop.ui.status) && (
-        <div>
-          <p className="font-bold text-white md:inline-block">Signed By: </p>
-          <p className="md:text-md text-sm font-bold text-white md:inline-block">
-            [ {[...prop.ui.signatures.keys()].join(", ")} ]
-          </p>
+const labelOfProposalContent = (content: proposalContent) => {
+  if ("changeThreshold" in content) {
+    return "Update threshold";
+  } else if ("adjustEffectivePeriod" in content) {
+    return "Update proposal duration";
+  } else if ("addOwners" in content) {
+    return `Add signer${content.addOwners.length > 1 ? "s" : ""}`;
+  } else if ("removeOwners" in content) {
+    return `Remove signer${content.removeOwners.length > 1 ? "s" : ""}`;
+  } else if ("transfer" in content) {
+    return `Transfer ${content.transfer.amount} mutez`;
+  } else if ("execute" in content) {
+    return "Execute";
+  } else if ("executeLambda" in content) {
+    return "Execute lambda";
+  }
+};
+
+type ProposalCardProps = {
+  id: number;
+  status: string;
+  date: Date;
+  activities: { signer: string; hasApproved: boolean }[];
+  content: proposalContent[];
+  proposer: { actor: string; timestamp: string };
+  resolver: { actor: string; timestamp: string } | undefined;
+  isSignable?: boolean;
+  isExpired?: boolean;
+  setCloseModal?: (arg: boolean | undefined) => void;
+};
+
+const ProposalCard = ({
+  id,
+  status,
+  date,
+  activities,
+  proposer,
+  resolver,
+  content,
+  isSignable = false,
+  isExpired = false,
+  setCloseModal,
+}: ProposalCardProps) => {
+  const [isOpen, setIsOpen] = useState(false);
+
+  const proposalDate = new Date(proposer.timestamp);
+  const resolveDate = new Date(resolver?.timestamp ?? 0);
+
+  return (
+    <div
+      className={`${
+        isOpen ? "h-auto" : isSignable ? "h-32" : "h-16"
+      } w-full overflow-hidden rounded bg-zinc-800 text-white`}
+    >
+      <button
+        className="grid h-16 w-full grid-cols-3 items-center gap-8 border-b border-zinc-900 px-6 py-4 lg:grid-cols-4"
+        onClick={() => {
+          setIsOpen(v => !v);
+        }}
+      >
+        <span className="justify-self-start font-bold">
+          <span className="mr-4 font-light text-zinc-500">
+            #{id.toString().padStart(2, "0")}
+          </span>
+          {status ?? "Rejected"}
+        </span>
+        <span
+          className="truncate font-light text-zinc-300"
+          style={{
+            minWidth: "7rem",
+          }}
+          title={content.map(labelOfProposalContent).join(", ")}
+        >
+          {content.map(labelOfProposalContent).join(", ")}
+        </span>
+        <span className="hidden min-w-max justify-self-end lg:block lg:translate-x-1/2">
+          {isSignable ? "Expires at: " : ""}
+          {date.toLocaleDateString()} -{" "}
+          {`${date.getHours().toString().padStart(2, "0")}:${date
+            .getMinutes()
+            .toString()
+            .padStart(2, "0")}`}
+        </span>
+
+        <div className="justify-self-end">
+          <TriangleDownIcon
+            className={`${isOpen ? "rotate-180" : ""} h-8 w-8`}
+          />
         </div>
-      )}
-      {"Proposing" === prop.ui.status && (
-        <div>
-          <p className="font-bold text-white md:inline-block">
-            Waiting for signatures from:{" "}
-          </p>
-          <p className="md:text-md text-sm font-bold text-white md:inline-block">
-            [
-            {signers(contract)
-              .filter(x => !!!prop.ui.signatures.find(p => x === p.signer))
-              .map(x => state.aliases[x] || x)
-              .join(", ")}{" "}
-            ]
-          </p>
-        </div>
-      )}
-      <div>
-        <p className="font-bold text-white md:inline-block">Transactions: </p>
-        <p className="md:text-md text-sm font-bold text-white md:inline-block">
-          [
-          {prop.ui.content
-            .map(x => `${renderContent(x, state, address, contract)}`)
-            .join(", ")}{" "}
-          ]
-        </p>
-      </div>
-      <div className="mt-4 flex flex-col md:flex-row">
-        <ContractLoader loading={loading}>
-          {state.address &&
-            signers(contract).includes(state.address) &&
-            signable && (
+      </button>
+      {isSignable && (
+        <div className="flex h-16 w-full items-center justify-around">
+          {!isExpired && (
+            <>
               <button
                 type="button"
                 className={
-                  !isOwner
-                    ? "pointer-events-none opacity-50 "
-                    : "" +
-                      "mx-none mx-auto block w-full self-center justify-self-end rounded bg-primary p-1.5 font-medium text-white hover:bg-red-500 hover:outline-none focus:bg-red-500 md:mx-auto md:w-1/3 md:self-end"
+                  "block w-full rounded bg-primary p-1.5 font-medium text-white hover:bg-red-500 hover:outline-none focus:bg-red-500 md:w-1/3"
                 }
                 onClick={async e => {
                   e.preventDefault();
-                  setCloseModal(false);
+                  setCloseModal?.(false);
                 }}
               >
                 Reject
               </button>
-            )}
-          {state.address &&
-            signers(contract).includes(state.address) &&
-            resolvable(prop.ui.signatures) &&
-            "Executed" !== prop.ui.status && (
               <button
                 type="button"
                 className={
-                  !isOwner
-                    ? "pointer-events-none opacity-50 "
-                    : "" +
-                      "mx-none mx-auto block w-full self-center justify-self-end bg-primary p-1.5 font-medium text-white hover:bg-red-500 hover:outline-none focus:bg-red-500 md:mx-auto md:w-1/3 md:self-end"
+                  "block w-full rounded bg-primary p-1.5 font-medium text-white hover:bg-red-500 hover:outline-none focus:bg-red-500 md:w-1/3"
                 }
                 onClick={async e => {
                   e.preventDefault();
-                  setCloseModal(undefined);
-                }}
-              >
-                Resolve
-              </button>
-            )}
-          {state.address &&
-            signers(contract).includes(state.address) &&
-            signable && (
-              <button
-                type="button"
-                className={
-                  "mx-none mx-auto block w-full self-center justify-self-end rounded bg-primary p-1.5 font-medium text-white hover:bg-red-500 hover:outline-none focus:bg-red-500 md:mx-auto md:w-1/3 md:self-end"
-                }
-                onClick={async e => {
-                  e.preventDefault();
-                  setCloseModal(true);
+                  setCloseModal?.(true);
                 }}
               >
                 Sign
               </button>
+            </>
+          )}
+
+          {isExpired && (
+            <button
+              type="button"
+              className={
+                "block w-full rounded bg-primary p-1.5 font-medium text-white hover:bg-red-500 hover:outline-none focus:bg-red-500 md:w-1/3"
+              }
+              onClick={async e => {
+                e.preventDefault();
+                setCloseModal?.(undefined);
+              }}
+            >
+              Resolve
+            </button>
+          )}
+        </div>
+      )}
+      <div></div>
+      <div className="space-y-4 px-6 py-4">
+        <section>
+          <span className="text-xl font-bold">Content</span>
+          <div className="mt-4 grid hidden w-full grid-cols-6 gap-4 text-zinc-500 lg:grid">
+            <span>Function</span>
+            <span className="flex items-center">
+              Metadata
+              <Tooltip text="Metadata is user defined. It may not reflect on behavior of lambda">
+                <InfoCircledIcon className="ml-2 h-4 w-4" />
+              </Tooltip>
+            </span>
+            <span className="justify-self-center">Amount</span>
+            <span className="justify-self-center">Address</span>
+            <span className="justify-self-end">Entrypoint</span>
+            <span className="justify-self-end">Parameters</span>
+          </div>
+          <div className="mt-2 space-y-4 font-light lg:space-y-2">
+            {content.map(renderProposalContent)}
+          </div>
+        </section>
+        <section>
+          <span className="text-xl font-bold">Activity</span>
+          <div className="mt-4 grid grid grid-cols-3 text-zinc-500">
+            <span>Date</span>
+            <span className="justify-self-center">Proposer</span>
+            <span className="justify-self-end">Status</span>
+          </div>
+          <div className="mt-2 space-y-2 font-light">
+            <div className="grid grid-cols-3">
+              <span className="w-full font-light">
+                {proposalDate.toLocaleDateString()} -{" "}
+                {`${proposalDate.getHours()}:${proposalDate.getMinutes()}`}
+              </span>
+              <span className="justify-self-center">
+                <Alias address={proposer.actor} />
+              </span>
+              <span className="justify-self-end">Proposed</span>
+            </div>
+            {activities.map(({ signer, hasApproved }, i) => (
+              <div key={i} className="grid grid-cols-3">
+                <span className="w-full justify-self-start font-light text-zinc-500">
+                  -
+                </span>
+                <span className="justify-self-center">
+                  <Alias address={signer} />
+                </span>
+                <span className="justify-self-end">
+                  {hasApproved ? "Approved" : "Rejected"}
+                </span>
+              </div>
+            ))}
+            {!!resolver && (
+              <div className="grid grid-cols-3">
+                <span className="justify-self-start font-light">
+                  {resolveDate.toLocaleDateString()} -{" "}
+                  {`${resolveDate.getHours()}:${resolveDate.getMinutes()}`}
+                </span>
+                <span className="justify-self-center">
+                  <Alias address={resolver.actor} />
+                </span>
+                <span className="justify-self-end">Resolved</span>
+              </div>
             )}
-          {state.address &&
-            signers(contract).includes(state.address) &&
-            !resolvable(prop.ui.signatures) &&
-            !signable &&
-            "Proposing" === prop.ui.status && (
-              <p className="mx-none mx-auto  block w-full self-center justify-self-end bg-primary p-1.5 font-medium text-white md:mx-auto md:w-1/3 md:self-end">
-                Waiting for signatures of other owners
-              </p>
-            )}
-        </ContractLoader>
+          </div>
+        </section>
       </div>
     </div>
   );
 };
 
-export default Card;
+// {
+//   state.address &&
+//     signers(contract).includes(state.address) &&
+//     resolvable(prop.ui.signatures) &&
+//     "Executed" !== prop.ui.status && (
+//       <button
+//         type="button"
+//         className={
+//           !isOwner
+//             ? "pointer-events-none opacity-50 "
+//             : "" +
+//               "mx-none mx-auto block w-full self-center justify-self-end bg-primary p-1.5 font-medium text-white hover:bg-red-500 hover:outline-none focus:bg-red-500 md:mx-auto md:w-1/3 md:self-end"
+//         }
+//         onClick={async e => {
+//           e.preventDefault();
+//           setCloseModal(undefined);
+//         }}
+//       >
+//         Resolve
+//       </button>
+//     );
+// }
+// {
+//   state.address && signers(contract).includes(state.address) && signable && (
+//     <button
+//       type="button"
+//       className={
+//         "mx-none mx-auto block w-full self-center justify-self-end rounded bg-primary p-1.5 font-medium text-white hover:bg-red-500 hover:outline-none focus:bg-red-500 md:mx-auto md:w-1/3 md:self-end"
+//       }
+//       onClick={async e => {
+//         e.preventDefault();
+//         setCloseModal(true);
+//       }}
+//     >
+//       Sign
+//     </button>
+//   );
+// }
+// {
+//   state.address &&
+//     signers(contract).includes(state.address) &&
+//     !resolvable(prop.ui.signatures) &&
+//     !signable &&
+//     "Proposing" === prop.ui.status && (
+//       <p className="mx-none mx-auto  block w-full self-center justify-self-end bg-primary p-1.5 font-medium text-white md:mx-auto md:w-1/3 md:self-end">
+//         Waiting for signatures of other owners
+//       </p>
+//     );
+// }
+export default ProposalCard;
