+++ pages/proposals.tsx
@@ -10,6 +10,7 @@ import fetchVersion from "../context/metadata";
 import { getProposals } from "../context/proposals";
 import { AppStateContext } from "../context/state";
 import { proposal, version } from "../types/display";
+import { countdown } from "../utils/adaptiveTime";
 import { getProposalsId, toProposal } from "../versioned/apis";
 
 const emptyProps: [number, { og: any; ui: proposal }][] = [];
@@ -76,6 +77,7 @@ const Proposals = () => {
   );
 
   const currentContract = state.currentContract ?? "";
+
   return (
     <div className="min-h-content relative flex grow flex-col">
       <Meta title={"Proposals"} />
@@ -122,25 +124,43 @@ const Proposals = () => {
             <div className="space-y-6">
               {filteredProposals
                 .sort((a, b) => b[0] - a[0])
-                .map(x => (
-                  <ProposalCard
-                    contract={state.contracts[currentContract]}
-                    id={x[0]}
-                    setCloseModal={(arg: boolean | undefined) =>
-                      setCloseModal({ proposal: [arg, x[0]], state: 4 })
-                    }
-                    key={JSON.stringify(x[1])}
-                    prop={x[1]}
-                    address={state.currentContract ?? ""}
-                    signable={
-                      !!state.address &&
-                      !!!x[1].ui.signatures.find(
-                        x => x.signer == state.address
-                      ) &&
-                      true
-                    }
-                  />
-                ))}
+                .map(x => {
+                  const deadline = new Date(
+                    new Date(x[1].ui.timestamp).getTime() +
+                      state.contracts[
+                        currentContract
+                      ]?.effective_period.toNumber() *
+                        1000
+                  );
+                  const hasDeadlinePassed = Date.now() >= deadline.getTime();
+
+                  return (
+                    <ProposalCard
+                      id={x[0]}
+                      key={x[0]}
+                      status={hasDeadlinePassed ? "Expired" : x[1].ui.status}
+                      date={deadline}
+                      activities={x[1].ui.signatures.map(
+                        ({ signer, result }) => ({
+                          hasApproved: result,
+                          signer,
+                        })
+                      )}
+                      content={x[1].ui.content}
+                      proposer={x[1].og.proposer}
+                      resolver={x[1].og.resolver}
+                      isSignable={
+                        !!state.address &&
+                        !!state.currentContract &&
+                        !x[1].ui.signatures.find(x => x.signer == state.address)
+                      }
+                      isExpired={hasDeadlinePassed}
+                      setCloseModal={arg => {
+                        setCloseModal({ proposal: [arg, x[0]], state: 4 });
+                      }}
+                    />
+                  );
+                })}
             </div>
           )}
         </div>
